fix(/build.mjs): fixes build failure on windows
Windows' environment variables may contain characters which are not dot-safe in JavaScript. This causes the build command to fail on Windows. Additionally, there is no chmod command on Windows, which introduces another failure.

This commit introduces shx as a dependency (to handle chmod absence) and introduces a `isDotSafeIdentifier` helper function to determine whether or not a key can be used as a dot-safe identifier. It assumes we can safely ignore any environment variables which seems to b a safe assumption.